* Fixed: correctly write item to guild bank item table in case swap  2 guild bank items.  Patch provided by rtfm. * Fixed: use int32 for store GetTotalAuraModifier internal accumulator, it can be negative in some cases. Patch provided by thenecromancer. * Explicitly check negative case in GetTotalAuraModifier call places.
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@5208 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/game/Unit.h
@@ -938,7 +938,7 @@ class MANGOS_DLL_SPEC Unit : public WorldObject
         AuraMap& GetAuras( ) {return m_Auras;}
         AuraList const& GetAurasByType(AuraType type) const { return m_modAuras[type]; }
         void ApplyAuraProcTriggerDamage(Aura* aura, bool apply);
-        long GetTotalAuraModifier(AuraType auratype) const;
+        int32 GetTotalAuraModifier(AuraType auratype) const;
         void SendMoveToPacket(float x, float y, float z, bool run, uint32 transitTime = 0);
         void setTransForm(uint32 spellid) { m_transform = spellid;}
         uint32 getTransForm() const { return m_transform;}
